Extract workflow model from ignite-worker module.
