--- grpc/workflow/model/src/main/java/org/opennms/poc/ignite/grpc/model/WorkflowTwin.java
@@ -1,16 +1,0 @@
-package org.opennms.poc.ignite.grpc.model;
-
-import java.util.UUID;
-
-public class WorkflowTwin {
-
-  private UUID id;
-
-  public UUID getId() {
-    return id;
-  }
-
-  public void setId(UUID id) {
-    this.id = id;
-  }
-}
