+++ grpc/workflow/publisher/src/main/java/org/opennms/poc/ignite/grpc/publisher/WorkflowPublisher.java
@@ -1,10 +1,10 @@
 package org.opennms.poc.ignite.grpc.publisher;
 
 import java.io.IOException;
-import org.opennms.poc.ignite.grpc.model.WorkflowTwin;
+import org.opennms.poc.ignite.model.workflows.Workflow;
 
 public interface WorkflowPublisher {
 
-  void publish(WorkflowTwin twin) throws IOException;
+  void publish(Workflow twin) throws IOException;
 
 }
