+++ ignite-worker/pom.xml
@@ -27,6 +27,12 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.opennms.poc</groupId>
+            <artifactId>model</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.apache.ignite</groupId>
             <artifactId>ignite-core</artifactId>
             <version>${ignite.version}</version>
